+++ CHANGELOG.md
@@ -1,15 +1,28 @@
 ChangeLog
 ---------
 
+### 1.20.2
+
+#### Core
+
+- **Update:** Fixed small memory leak.
+- **Update:** Fixed the detail view bug with the `td` instead of `icon`.
+
+#### Extensions
+
+- **Update(export):** Fixed XSS vulnerability bug by onCellHtmlData.
+- **Update(export):** Fixed export footer bug without setting height.
+- **Update(filter-control):** Fixed the comparison of dates when using the `datepicker`.
+
 ### 1.20.1
 
-### Core
+#### Core
 
 - **Update:** Fixed toggle column bug with complex headers.
 - **Update:** Fixed icons option cannot work bug when it's a string.
 - **Update:** Updated TypeScript definitions.
 
-##### Extensions
+#### Extensions
 
 - **Update(cookie):** Fixed cookie extension error with multiple-sort.
 - **Update(export):** Fixed the `exportOptions` option cannot support the data attribute.
@@ -36,7 +49,7 @@ ChangeLog
 - **Update:** Fixed `onVirtualScroll` not define default method.
 - **Update:** Updated cs-CZ, ko-KR, nl-NL, nl-BE, bg-BG, fr-LU locales.
 
-##### Extensions
+#### Extensions
 
 - **New(filter-control):** New version of filter-control with new features.
 - **New(reorder-rows):**: Added `onAllowDrop` and `onDragStop` options.