+++ site/news.md
@@ -4,17 +4,32 @@ title: News
 description: News and announcements for all things Bootstrap Table, including new releases.
 ---
 
+## Bootstrap Table 1.20.2
+
+<span class="post-date">25 May 2022</span>
+
+#### Core
+
+- **Update:** Fixed small memory leak.
+- **Update:** Fixed the detail view bug with the `td` instead of `icon`.
+
+#### Extensions
+
+- **Update(export):** Fixed XSS vulnerability bug by onCellHtmlData.
+- **Update(export):** Fixed export footer bug without setting height.
+- **Update(filter-control):** Fixed the comparison of dates when using the `datepicker`.
+
 ## Bootstrap Table 1.20.1
 
 <span class="post-date">12 May 2022</span>
 
-### Core
+#### Core
 
 - **Update:** Fixed toggle column bug with complex headers.
 - **Update:** Fixed icons option cannot work bug when it's a string.
 - **Update:** Updated TypeScript definitions.
 
-##### Extensions
+#### Extensions
 
 - **Update(cookie):** Fixed cookie extension error with multiple-sort.
 - **Update(export):** Fixed the `exportOptions` option cannot support the data attribute.
@@ -42,7 +57,7 @@ description: News and announcements for all things Bootstrap Table, including ne
 - **Update:** Fixed `onVirtualScroll` not define default method.
 - **Update:** Updated cs-CZ, ko-KR, nl-NL, nl-BE, bg-BG, fr-LU locales.
 
-##### Extensions
+#### Extensions
 
 - **New(filter-control):** New version of filter-control with new features.
 - **New(reorder-rows):**: Added `onAllowDrop` and `onDragStop` options.